refactor: move stats to widgets layer
