+++ src/widgets/stats/index.ts
@@ -1,0 +1,3 @@
+import RoundStats from "./ui/RoundStats";
+
+export default RoundStats;
