+++ src/widgets/stats/ui/ArtworkMiniature.module.css
