+++ src/widgets/stats/ui/RoundStats.ts
@@ -2,9 +2,11 @@ import { Component, Button, Modal, div, i } from "../../../shared/ui";
 import ArtworkMiniature from "./ArtworkMiniature";
 import SentencesList, { SentencesListType } from "./SentencesList";
 
-import RoundState, { type Stage } from "../model/RoundState";
+import RoundState, {
+  type Stage,
+} from "../../../features/game/model/RoundState";
 import { Observer, Publisher } from "../../../entities/state";
-import { StageStatus } from "../model/StageStatus";
+import { StageStatus } from "../../../features/game/model/StageStatus";
 
 import styles from "./RoundStats.module.css";
