+++ src/widgets/stats/ui/SentencesList.module.css
